double to float

+++ AAI-Final-Assignment-WinForms/Entities/Projectile.cs
@@ -10,8 +10,8 @@ namespace AAI_Final_Assignment_WinForms.Entities
 {
     public class Projectile : BaseGameEntity
     {
-        public Projectile(Vector2D pos, GameWorld world, double scale, int textureWidth, int textureHeight,
-            double radius) : base(pos, world, scale, textureWidth, textureHeight, radius) {
+        public Projectile(Vector2D pos, GameWorld world, float scale, int textureWidth, int textureHeight,
+            float radius) : base(pos, world, scale, textureWidth, textureHeight, radius) {
 
         }
     }
